requirements.txt
flask
flask-wtf
requests
python-dotenv
gunicorn
black

